Adding guest checkout (#69)
* working post route in cart.js to accomodate user and guest ability to add to cart

* created cart wrapper component to conditionally render cart for user vs guest; all cart routes able to handle guest and user checkout and payment

* resolved comments from PR

+++ client/components/checkout.js
@@ -32,7 +32,7 @@ class DisconnectedCheckout extends React.Component {
     })
 
     axios.post('/api/orders', {orderItems: orders})
-    axios.delete('/api/cart', {data: {userId: this.props.user.id}})
+    if(this.props.user) axios.delete('/api/cart', {data: {userId: this.props.user.id}})
     this.setState({cartItems: []})
   }
